requirements.txt
pyserial
numpy
psutil
evdev
pynput
pyyaml
pyinstaller
requests >= 2.32.5
python-iplocate >= 1.0.0
dotenv >= 1.0.0
